+++ sphinx/source/changelog.rst
@@ -6,8 +6,8 @@ Changelog
 * Changed the default value of the ``accelerate-dec-a`` simulated LOAD
   configuration parameter for :ref:`tap2sna.py <tap2sna-sim-load>` from ``1``
   to ``3``
-* Added the ``activision``, ``codemasters`` and ``diver`` tape-sampling loop
-  :ref:`accelerators <tap2sna-accelerators>`
+* Added the ``activision``, ``codemasters``, ``diver`` and ``mirrorsoft2``
+  tape-sampling loop :ref:`accelerators <tap2sna-accelerators>`
 
 9.5 (2025-01-28)
 ----------------